Filling in the methods
Former-commit-id: 529be4f9ae069281068984d6cacc7f7aa7a1a0ae

+++ Echo.Core/src/Aggregation/Primitives/LightType.cs
@@ -2,7 +2,12 @@
 
 namespace Echo.Core.Aggregation.Primitives;
 
-public enum LightType : byte
+/// <summary>
+/// The light type of an <see cref="EntityToken"/> if that <see cref="EntityToken"/> is of type <see cref="TokenType.Light"/>.
+/// </summary>
+/// <remarks>Because of the data size <see cref="EntityToken"/>, at a maximum there can only be 64
+/// different <see cref="LightType"/>. See <see cref="EntityToken.LightIndexCount"/> for more.</remarks>
+public enum LightType : uint
 {
 	/// <summary>
 	/// Represents an <see cref="InfiniteLight"/>.
